backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain_community
llama-cpp-python